requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.3.2
pydantic==2.6.4
python-dotenv==1.0.0